[PL-124] Feat: 플랜 전체보기 레이아웃 정돈
