+++ lib/ui/common/comopnent/planit_button.dart
@@ -54,7 +54,7 @@ class PlanitButton extends StatelessWidget {
             ? PlanitColors.black01
             : PlanitColors.black01.withValues(alpha: 0.5))
         : ((buttonColor == PlanitButtonColor.white)
-            ? PlanitColors.white01
+            ? PlanitColors.white03
             : PlanitColors.red);
 
     final Color overlayColor = (buttonColor == PlanitButtonColor.black)
